fix(control-plane): retain agent context in status history
